+++ packages/common/src/main/java/io/github/trethore/graphene/internal/cef/GrapheneCefDirectoryPickerIntentSource.java
@@ -1,0 +1,5 @@
+package io.github.trethore.graphene.internal.cef;
+
+interface GrapheneCefDirectoryPickerIntentSource {
+    boolean consumeDirectoryPickerIntent();
+}